Prototype now works flowlessly
* corrected some runtime errors that accured on the server side based on messeges being sent with the wrong data type
* corrected some flawed connections and attributes in a couple of classes that were not defined correctly
* might need to be updated in the future !!!

+++ hibernate_module/src/main/java/il/cshaifasweng/Interfaces/DAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface DAO<T> {
     SessionFactory factory =MySQL.getSessionFactory();
-    T get(long id,Class<T> type);
+    T get(int id,Class<T> type);
     List<T> getAll(Class<T> type);
     void save(T t,Class<T> type);
     void update(T t);
